+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 3.3.3 (TBD)
 
-* TBD
+* GeoJsonObject simple geometry retrieval method
+* FeatureConverter conversion to simple geometry utility methods
 
 ## [3.3.2](https://github.com/ngageoint/simple-features-geojson-java/releases/tag/3.3.2) (01-19-2023)